pyproject.toml
[project]
name = "rt-process"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "channels>=4.3.1",
    "daphne>=4.2.1",
    "django==6.0a1",
    "pillow>=11.3.0",
    "redis>=6.4.0",
    "django-tasks>=0.8.1",
    "channels-redis>=4.3.0",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
]
